Added Group Collision option to Collision Sensor
works great for our game jam/example game but might be too ambigous. Takes a vector of Sprites so now you can compare one sprite with a list of sprites with the intention to stop checking as soon as one is hit. That isn't ambigous but we only activate the acturator that is indexed with the sprite detected.